more updates and stealing Darks 049 update

+++ code/datums/actions/scp/scp049cure.dm
@@ -7,7 +7,7 @@
 	button_icon = 'icons/scp/scp-049.dmi'
 	button_icon_state = "cure"
 	background_icon_state = "bg_default"
-	
+
 	/// Range for cure action
 	var/cure_range = 2
 	/// Cooldown for cure action
@@ -17,24 +17,28 @@
 
 /datum/action/scp049cure/New(Target)
 	..()
-	if(!istype(owner, /mob/living/carbon/human/scp049))
+	// Don't delete if owner is not set yet - it will be set when granted
+	if(owner && !istype(owner, /mob/living/carbon/human/scp049))
 		qdel(src)
 
 /datum/action/scp049cure/IsAvailable(feedback = FALSE)
 	if(!..())
 		return FALSE
-	
+
+	if(!owner || !istype(owner, /mob/living/carbon/human/scp049))
+		return FALSE
+
 	if(world.time < cure_cooldown)
 		if(feedback)
 			to_chat(owner, "<span class='warning'>The cure needs time to prepare...</span>")
 		return FALSE
-	
+
 	// Check if there are any valid targets in range
 	var/mob/living/carbon/human/scp049/scp = owner
 	for(var/mob/living/carbon/human/H in range(cure_range, scp))
 		if(HAS_TRAIT(H, TRAIT_PESTILENCE))
 			return TRUE
-	
+
 	if(feedback)
 		to_chat(owner, "<span class='warning'>No afflicted targets in range.</span>")
 	return FALSE
@@ -42,56 +46,62 @@
 /datum/action/scp049cure/Trigger(trigger_flags)
 	if(!..())
 		return FALSE
-	
+
 	if(!IsAvailable(TRUE))
 		return FALSE
-	
+
+	if(!owner || !istype(owner, /mob/living/carbon/human/scp049))
+		return FALSE
+
 	var/mob/living/carbon/human/scp049/scp = owner
-	
+
 	// Get list of valid targets
 	var/list/targets = list()
 	for(var/mob/living/carbon/human/H in range(cure_range, scp))
 		if(HAS_TRAIT(H, TRAIT_PESTILENCE))
 			targets += H
-	
+
 	if(!length(targets))
 		to_chat(scp, "<span class='warning'>No afflicted targets in range.</span>")
 		return FALSE
-	
+
 	var/mob/living/carbon/human/target
 	if(length(targets) == 1)
 		target = targets[1]
 	else
 		target = input(scp, "Select target to cure:", "Cure Target") as null|mob in targets
-	
+
 	if(!target)
 		return FALSE
-	
+
 	if(get_dist(scp, target) > cure_range)
 		to_chat(scp, "<span class='warning'>Target is too far away.</span>")
 		return FALSE
-	
+
 	if(!HAS_TRAIT(target, TRAIT_PESTILENCE))
 		to_chat(scp, "<span class='warning'>Target is not afflicted with pestilence.</span>")
 		return FALSE
-	
+
 	// Perform cure
 	perform_cure(scp, target)
 	return TRUE
 
 /datum/action/scp049cure/proc/perform_cure(mob/living/carbon/human/scp049/scp, mob/living/carbon/human/target)
+	if(!scp || !target) // Safety check
+		return
+
 	cure_cooldown = world.time + cure_cooldown_time
-	
-	// Update button availability
-	UpdateButtons()
-	
+
 	// Call the cure function
 	scp.cure_target(target)
 
-/datum/action/scp049cure/UpdateButtons()
-	..()
+/datum/action/scp049cure/proc/update_cure_button()
 	// Update button based on availability
+	if(!owner) // Safety check
+		return
+
 	if(IsAvailable())
 		button_icon_state = "cure"
 	else
 		button_icon_state = "cure_disabled"
+	build_all_button_icons(UPDATE_BUTTON_ICON, FALSE)
